libro digital

+++ Programacion/Tarea_u6/src/Libreria/Descargable.java
@@ -10,7 +10,7 @@ package Libreria;
  */
 public interface Descargable {
     
-    double getSize();
+    int getSize();
     int getNumDescargas();
     int descargar();
